fix(dynamic-secrets): normalize dynamic-secret name and reject empty usernames
