+++ backend/src/ee/services/dynamic-secret/providers/templateUtils.test.ts
@@ -1,6 +1,8 @@
 import { describe, expect, test } from "vitest";
 
-import { compileUsernameTemplate } from "./templateUtils";
+import { TDynamicSecrets } from "@app/db/schemas/dynamic-secrets";
+
+import { compileUsernameTemplate, toSafeUsername } from "./templateUtils";
 
 describe("compileUsernameTemplate", () => {
   test("renders the identity name when it is a plain identifier", () => {
@@ -39,6 +41,18 @@ describe("compileUsernameTemplate", () => {
         unixTimestamp: 1700000000
       })
     ).toBe("dynamic-secret-sa-abc123");
+  });
+
+  test("normalizes the dynamic secret name used in a template", () => {
+    expect(
+      compileUsernameTemplate({
+        usernameTemplate: "{{dynamicSecret.name}}",
+        randomUsername: "unused",
+        identity: null,
+        dynamicSecret: { name: "my secret", type: "sql-database" } as unknown as TDynamicSecrets,
+        unixTimestamp: 1700000000
+      })
+    ).toBe("my_secret");
   });
 
   test("applies the toUpperCase option", () => {
@@ -82,5 +96,40 @@ describe("compileUsernameTemplate", () => {
         unixTimestamp: 1700000000
       })
     ).toThrow("unsupported characters");
+  });
+
+  test("rejects an empty compiled username", () => {
+    expect(() =>
+      compileUsernameTemplate({
+        usernameTemplate: "{{random 0}}",
+        randomUsername: "unused",
+        identity: null,
+        dynamicSecret: null,
+        unixTimestamp: 1700000000
+      })
+    ).toThrow("unsupported characters");
+  });
+});
+
+describe("toSafeUsername", () => {
+  test("leaves a plain identifier unchanged", () => {
+    expect(toSafeUsername("build_agent01")).toBe("build_agent01");
+  });
+
+  test("replaces disallowed characters with underscore", () => {
+    expect(toSafeUsername("name with spaces")).toBe("name_with_spaces");
+    expect(toSafeUsername("a.b@c")).toBe("a_b_c");
+  });
+
+  test("keeps letters, digits, underscore and hyphen", () => {
+    expect(toSafeUsername("ci-runner_01")).toBe("ci-runner_01");
+  });
+
+  test("caps the result at 63 characters", () => {
+    expect(toSafeUsername("a".repeat(100))).toHaveLength(63);
+  });
+
+  test("falls back to a default when nothing remains", () => {
+    expect(toSafeUsername("")).toBe("inf_user");
   });
 });
